refactor: code splitting for lambda function
